Quick fix to make annual report work + automatic tweet "shrinker"
First month of the year is 1, not 0, now annual report is actually
launched when script is ran in January.

Automatic quick tweet shortener when 280 characters is exceeded. At the
moment, three templates are provided for the shortener (normal, short,
and shorter) and the longest one that fits in the space is used.
Probably more work should be done in the future to improve the
templates, and provide an extra mechanism to even shorten up album
titles when they are really long.

Some code clean-up + reorganization.