Project renamed + now compatible with Mastodon
Project now is able to publish messages to Mastodon and Twitter. The
project name has been accordingly renamed to reflect that change.

Now, when credentials for Twitter or Mastodon are missing, the
functionality will be completely disabled for it, so the output of the
program will be cleaner, and it'll run faster (since it won't try to
submit the message to a non-working service multiple times with extra
delay time between retries.